fixing after review
Required
a modal should pop up to congratulate the player if all cards have been matched, you could make use of :link: SweetAlert2 as a dependency

Required
The star rating works well, but the star rating cannot be lower than 1 star, the lowest should be 1 star

Required
When the player starts a game, a displayed timer should also start. Once the player wins the game, the timer stops.

Required
The README.md file should be updated or re-written from scratch, make sure to remove everything related to the prior starter project, and it should include all the dependencies used to build the game such as font-awesome and SweetAlert2 if used